Rename e2esetup/helm to e2esetup/chart
