+++ integration/test/migration/migration_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/giantswarm/e2e-harness/pkg/framework"
-	"github.com/giantswarm/e2esetup/helm/env"
+	"github.com/giantswarm/e2esetup/chart/env"
 	"github.com/giantswarm/microerror"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
